server/m5rs/pyproject.toml
[project]
name = "m5rs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "William Tumeo", email = "git@tumeo.space" }]
requires-python = ">=3.12"
dependencies = ["pillow>=11.2.1", "python-xlib>=0.33", "websockets>=15.0.1"]

[project.scripts]
m5rs = "m5rs:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
